Fix memory leaks in OCaml bindings
To fix memory leaks, we use the finalise option from the allocate call
of the ctypes library instead of the Gc call.

For values that are fully converted to OCaml, if they require a
finalisation function to be called when we unwrap them, we call it after
its use.

TN: S916-052